# Use memcpy() to dump true/false/null